requirements.txt
imagesize
opencv-python
pillow
python-dateutil
rich
rich_argparse
shtab
tqdm
cfg-argparser
imagehash
autopep8
polars
typer[all]
pyarrow